refactor: Base Models interface Enhancement (Carlos Pimentel)
The base Models interface from models was refactor to be
more flexible and now the UserRepository or PostRepository
or XModelRepository could receive as a concrete type any
model that had been previously implemented the Models
interface, in this current proposal the User Model does that.

That would make a great improvement for the app making it
more scalable than what already is which only receives
the User Model as an argument to its UserRepository.